requirements.txt
playwright
sqlalchemy
apscheduler
slack_sdk
beautifulsoup4
fake-useragent
aiohttp
slack_bolt
python-dotenv
